+++ models/recipe.go
@@ -1,0 +1,14 @@
+package models
+
+import (
+	"time"
+)
+
+type Recipe struct {
+	ID           string    `json:"id"`
+	Name         string    `json:"name"`
+	Tags         []string  `json:"tags"`
+	Ingredients  []string  `json:"ingredients"`
+	Instructions []string  `json:"instructions"`
+	PublishedAt  time.Time `json:"publishedAt"`
+}
