+++ swagger.yaml
@@ -1,0 +1,2 @@
+paths: {}
+swagger: "2.0"
